refactor: 코드 리뷰 반영 Alert Extension 형식 변경
action을 외부에서 주입받도록 변경
