+++ ios-contact-manager/Support/UIViewController+Extension.swift
@@ -1,0 +1,17 @@
+import UIKit
+
+extension UIViewController {
+    
+    func showMessageAlert(title: String? = nil,
+                                   message: String,
+                                   action: [UIAlertAction],
+                                   completion: (() -> ())? = nil) {
+        let alertViewController = UIAlertController(title: title, message: message, preferredStyle: .alert)
+        
+        for result in action {
+            alertViewController.addAction(result)
+        }
+        
+        self.present(alertViewController, animated: true, completion: completion)
+    }
+}
